feat: random JWT secrets on wallet daemon and signaling server (#682)
Description
---
* Add a new common utility function on `tari_common_types` for secret
generation
* Update wallet daemon and signaling server JWT secrets to be randomized
at startup

Motivation and Context
---
Right now JWT secrets are hardcoded both in the wallet daemon as well as
in the signaling server. This is not acceptable from a security
standpoint if we want to create a public test network. So the purpose of
this PR is to change both the wallet daemon and the signaling server to
haver random JWT secrets each time they start.

Please do not merge this PR until
https://github.com/tari-project/tari-connector/pull/9 is merged

How Has This Been Tested?
---
Manually launching both applications and inspecting the JWTs

What process can a PR reviewer use to test or verify this change?
---
Add log instructions to see the JWTs and launch both wallet daemon and
signaling server

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify