fix: indexer webui mime (#722)
Description
---
The content type in header was not set for webui.

Motivation and Context
---
The ui will not render properly when the content type is not set.

How Has This Been Tested?
---
Manually checking the content type in header in indexer webui.

What process can a PR reviewer use to test or verify this change?
---
Same as above.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify