feat: cache balance merkle tree (#707)
Description
---
Cache the BMT per epoch.

Motivation and Context
---
It was computed everytime we needed it, but it's a constant for the
whole epoch.

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Just run the VN with some transaction and you will see that the BMT is
stored in the DB.


Breaking Changes
---

- [x] None