feat: add jrpc address for webui to connect to (#738)
Description
---

Motivation and Context
---

How Has This Been Tested?
---
I created a new
[PR](https://github.com/tari-project/dan-testing/pull/43) in dan-testing
that set the listening address to `0.0.0.0:<port>` and the webui jrpc
address to `127.0.0.1:<port>`.

What process can a PR reviewer use to test or verify this change?
---
Either do the same as me or test my other 


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify