feat(indexer): new jrpc method to get all transactions related to a substate (#737)
Description
---
Added new jrpc method `get_substate_transactions` that returns a
(ordered) list of transactions that updated the requested substate
address.

Motivation and Context
---
For the standalone Metamask Snap wallet we want to show the list of
transactions. As the snap is standalone (no wallet daemon connection),
we need an easy way to fetch all the transactions of a particular
account (component substate) from the indexer. This PR adds a new method
to accomplish this, in a general way that can be used for any kind of
substate address (component, vaults, nfts, etc.).

How Has This Been Tested?
---
Manually querying the new method using Postman

What process can a PR reviewer use to test or verify this change?
---
Query the new method using Postman

Breaking Changes
---
- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify