refactor: token symbol should not have to be unique per template (#743)
Description
---
Changed the resource address calculation to derive them from the
transaction hash instead of the symbol

Motivation and Context
---
Right now the resource address symbol must be unique among all the
components of a template address. This imposes a restriction when
developing templates.

This PR refactors resource addresses from being calculated as
`H(template_address || symbol)` to being calculated as `H(tx_hash || n)`
where `n` is an auto-increment index.

Just for context, the original PR
https://github.com/tari-project/tari-dan/pull/450 that added the symbol
as part of the resource address did it to solve the problem of having to
specify the newly created outputs shards in the transaction. This was
mostly happening when minting NFTs in the same transaction that created
the parent resource. But we no longer require to specify the new outputs
to be created in a transactions, so we can safely revert to use default
auto-increment ids.

Also, there is an old draft PR
https://github.com/tari-project/tari-dan/pull/529 that tried to address
this problem and can be abandoned if this PR is approved and merged.

How Has This Been Tested?
---
New cucumber scenario that creates a non fungible resource and mints
NFTs in the same transaction.

What process can a PR reviewer use to test or verify this change?
---
Try to create a resource and mint an NFT in the same transaction

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify