chore: remove unused transaction fields (#747)
Description
---
Removed all unused transactions fields including all references to them
from anywhere in the project

Motivation and Context
---
There are some fields that are currently ignored when submitting a
transaction:
* `new_resources`
* `specific_non_fungible_outputs`
* `non_fungible_mint_outputs`
* `new_non_fungible_outputs`
* `new_non_fungible_index_outputs`

In the past they were needed to be able to fill the new output shards
when building the transaction, but this is no longer necessary as the VN
is able to infer new outputs when processing the transaction.

The purpose of this PR is to remove all these unused fields, alongside
any component that used them (clients, CLI commands, test utils, etc).

How Has This Been Tested?
---
Existing unit and integration tests pass

What process can a PR reviewer use to test or verify this change?
---
Nothing specific, behaviour in the project should remain the same

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify