feat: create registration file on startup (#752)
Description
---
Creates a registration file on startup. 
Adds a fee claim public key to the config file 

Motivation and Context
---
Currently, the only way to register a validator node is by calling the
json-rpc, via CLI or the web UI. This allows a user to register the
validator node via an unconnected minotari wallet.

I've left the jsonrpc registration as is for now, but I think it should
be removed as it is unsafe to leave a hot wallet running on the
validator node

How Has This Been Tested?
---
Tested start up.

What process can a PR reviewer use to test or verify this change?
---
Run the VN, look for the registration.json file in the base path

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify