fix(template)!: make token_symbol part of the metadata again in resources (#746)
Description
---
Integrated back the `token_symbol` field into the `metadata` field in
resources

Motivation and Context
---
After https://github.com/tari-project/tari-dan/pull/743, the
`token_symbol` field in resources is not used anymore for address
calculation. This means that the original reason the field was made
mandatory thought the whole project (and extracted out of the `metadata`
field) is no longer present.

This PR makes the token `token_symbol` field in resources optional
again, by integrating it into the `metadata` field as originally
implemented.

How Has This Been Tested?
---
* All existing unit and integration tests pass
* Manually started a fresh new network, created an account with funds
and inspected the balances in the wallet daemon

What process can a PR reviewer use to test or verify this change?
---
Launch the network and perform operations on resources

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Requires a full network reset, as the template engine API to
create resources changes, as well as the account_nfts builtin template