feat: adds settings JSON-RPC call to allow changes to indexer URL (#744)
Description
---
Add new jrpc calls that will allow change of the indexer jrpc url in the
wallet daemon.

Motivation and Context
---

How Has This Been Tested?
---
as below

What process can a PR reviewer use to test or verify this change?
---
Run dan-testing with 2 indexers. And call the jrpc method to change it.
And e.g. send money using the new indexer. Then you should clearly see
in the logs which one was used.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify