feat: add metadata string parser (#780)
Description
---
Adds the ability to pass in a string that will converted to a Metadata
table in the dan_wallet_client

Motivation and Context
---
It was previously not possible to create a Metadata object from args
passed to a scaffolded CLI program. I believe this would have been a
problem for JS clients as well.

This makes it possible to pass a CLI arg in the form `key=value,
key2=value` and it will be converted to a Metadata table.

How Has This Been Tested?
---
Tested using the scaffold of the stablecoin contract, passing in
`provider_name=mike`

What process can a PR reviewer use to test or verify this change?
---
Scaffold a contract that takes in a parameter of type Metadata, then
call that method using the above syntax


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify