fix: multi committee transaction (#768)
Description
---
This fixes two bugs:
- every VN now distributes the transaction after execution (we can maybe
limit this to the VNs that have the TxReceipt, because that's the shard
where is it always distributed).
- VNs is no longer trying to delete substates that are in foreign
committees.

Motivation and Context
---

How Has This Been Tested?
---
Running dan-testing with 2 committees and several transactions.

What process can a PR reviewer use to test or verify this change?
---
Same as above.


Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Older VNs are currently not working 100% properly and they
need to be recompiled.