fix: the genesis block_id values should be 0 (#786)
Description
---
Sometimes the real hash of genesis block was used. Which resulted in a
queries for a block that never existed because the real block_id of
genesis is not zero.

Motivation and Context
---

How Has This Been Tested?
---
Letting the github actions do it.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify