feat: foreign proposal command (#792)
Description
---
It's build on top of #757 
In current state everyone propose the foreign proposal command all the
time.
If I build on top of block where the foreign proposal is, then I can not
propose it and mark it as proposed.
But everyone else needs to check if there is a block on top of the block
where the command is to mark it as well. But the command is not in the
locked block. So it can be rollbacked, right?
For the transactions we have something similar where we keep track of
the changes until they are locked, in the
`transaction_pool_state_updates` table.

Motivation and Context
---

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify