feat: set refresh balance interval to 5 sec (#819)
Description
---
Set refresh balance interval to 5 sec (was manual only when the balance
redraw was triggered). 5sec is the same as for the transaction.

Motivation and Context
---

How Has This Been Tested?
---
Same as below.

What process can a PR reviewer use to test or verify this change?
---
Setup a dan wallet. Send some money. Wait for UI change.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify