docs(template_lib): add rustdoc documentation (#824)
Description
---
* Adds rustdoc to all the relevant elements present in the crate.
* Adding a convenience method `with_token_symbol` to fungible tokens,
similar to the one present in other types of resources.

Motivation and Context
---
The `template_lib` crate is the main API that template developers will
use on a day-to-day basis to develop Tari templates. It's important for
it to have extensive and useful documentation.

This PR adds rustdoc to all the relevant elements present in the
`template_lib` crate.

Ref https://github.com/tari-project/tari-dan/issues/336

How Has This Been Tested?
---
Not applicable

What process can a PR reviewer use to test or verify this change?
---
Run `cargo doc` and inspect the resulting documentation

Breaking Changes
---
- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify