feat(integration_tests): new test for downed substates (#798)
Description
---
* Refactored the VN util of the integration test so method calls return
a result, making it easier to assert expected errors.
* New cucumber test for downed local substates.

Motivation and Context
---
We want to assert that a substate that has been downed cannot be used
again as an input. There are two situations:
* The substate is local to the VN
* The substate is a remote committee

This PR adds a new test for the first situation, downed local substates.
The second situation requires a more advance integration test setup
where we can manipulate multiple committees deterministically, so it's
left out of scope of this PR.

Ref https://github.com/tari-project/tari-dan/issues/387

How Has This Been Tested?
---
New test passes

What process can a PR reviewer use to test or verify this change?
---
Execute the test

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify