feat: disable buttons on send, add result dialog (#813)
Description
---
Disable buttons in send dialog when you press send. Add result dialog to
the ui.
This dialog can be used from anywhere for any purpose not just sending
result.
**The transaction is successful even when only fees are accepted !**

Motivation and Context
---
It was very annoying not knowing if you pressed the button to send or
not.

Closes #814 

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Send some funds and see the result or fail, depending on what you enter.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: stringhandler <mikethetike@tari.com>