feat: sort TXs in UI, add timestamp (#804)
Description
---
The sort is done on DB level. But I added the timestamp to output anyway
so we can see it in the UI.

Motivation and Context
---

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Run the wallet with some TXs to see the changes.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

Co-authored-by: stringhandler <mikethetike@tari.com>