feat: do fee estimation in UI transfer (#826)
Description
---
Add fee estimation to transfer in the UI.
The way it works, the fee field is readonly, instead of "send tari"
there is a "Estimate Fee", once you click it, it will calculate the fees
and populate the fee field and the button changes to "Send", if you
change anything, it will go back to "Estimate Fee".
The problem is that we have to enter the maximum fee for estimation. If
user don't have enough money (the amount for transfer+the maximum fee
for calculation) that's problem. In current state the maximum fee is set
to 1000. So if it's going to be more, the estimation.

Motivation and Context
---

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Go to the send dialog in the wallet ui.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify