feat: create get_non_fungible_ids action for buckets (#847)
Description
---
* Created a new `GetNonFungibleIds` action for buckets in the engine
* Created a `get_non_fungible_ids` method in `template_lib` for Buckets,
that calls the new engine action

Motivation and Context
---
Currently in `template_lib` we have a convenient `get_non_fungible_ids`
method in vaults that returns all the NFT ids contained in the vault.
Buckets should also have the same method, so this PR adds it.

How Has This Been Tested?
---
Manually with a template that uses the new method

What process can a PR reviewer use to test or verify this change?
---
Create a template that uses the new method

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify