chore(engine): remove unused NFT marketplace POC template (#854)
Description
---
* Deleted the files of the NFT marketplace test template in the `engine`
crate

Motivation and Context
---
Up until now, we had an unused test template inside the `engine` crate.
It was created as a proof of concept for a NFT marketplace, but it never
compiled nor had tests because it needed features (like composability)
that were not present back then.

We now have a repository for template examples that include the updated
version of the NFT marketplace
(https://github.com/tari-project/test-templates/pull/2), so this PR
simply removes the old proof of concept template from the `engine`
crate.

How Has This Been Tested?
---
Testing does not apply

What process can a PR reviewer use to test or verify this change?
---
Does not apply

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>