feat(template_lib): add from_str for NFT addresses (#855)
Description
---
* Add from_str string for NFT addresses
* Refactor the `SubstateAddress` parsing for NFTs so it reuses the new
parsing

Motivation and Context
---
In template code we need a convenience method to parse a
`NonFungibleAddress`, because right now the only way is to separately do
it for the resource address, then for the NFT id and at the end build
the address.

How Has This Been Tested?
---
* New unit tests for `NonFungibleAddress` parsing
* Existing unit tests for `SubstateAddress` pass

What process can a PR reviewer use to test or verify this change?
---
Try to parse a NFT address from inside template code.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify