feat: sign proposal (#858)
Description
---
Sing the proposal (the proposer signs the block).

Motivation and Context
---

How Has This Been Tested?
---
I just did sample run on dan-testing.

What process can a PR reviewer use to test or verify this change?
---
You can run the dan-testing and maybe do a debug logs on the signature
check to make sure that the checks are enforced.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify