fix: claim burn to create account (#848)
Description
---
The claim burn was working only for account that were already created. I
refactored the code for create-free-test-coins as well, because it has a
lot in common with claiming burns.

Motivation and Context
---
Before this PR there was no way to create an account without having fees
and claiming burn needed an account to work.

How Has This Been Tested?
---
Burned and claimed funds for a new account.

What process can a PR reviewer use to test or verify this change?
---
The claim public key for burning needs to be filled with key from the
dan wallet. After that you can call the claim burn with new account name
and key index for the respective key.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>