test: break infinity loop in template (#866)
Description
---
Test that the infinity loop will eventually break because of the
increasing costs.

Motivation and Context
---

How Has This Been Tested?
---
There is a new test `test_break_infinity_loop`

What process can a PR reviewer use to test or verify this change?
---
The cost for loop was increased to 1 (from 0), so we should confirm that
nothing else is breaking.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify