feat(template_lib): builtin template addresses (#881)
Description
---
* `template_lib` exports an enum `BuiltinTemplate` that enumerates all
the builtin templates and allows to retrieve the addresses via a new
`BuiltinTemplateInvoke` engine operation
* The engine implements the `EngineOp::BuiltinTemplateInvoke` by
accessing the appropriate constants in the `template_builtin` crate

Motivation and Context
---
Template developers will want to perform cross-template calls with
builtin templates (e.g. accounts). For convenience we want to expose the
corresponding template addresses in the `template_lib` crate.

There are two main ways of doing this:

1. Export the template addresses directly as constants
2. Create a new engine call for getting the template address

This PR follows (2) for template WASM size optimisation reasons, as (1)
would require to include the full address (32 bytes for each one of the
builtin templates, and we may have more of them in the future) in every
template WASM.

How Has This Been Tested?
---
New engine unit test for builtin template addresses

What process can a PR reviewer use to test or verify this change?
---
Use the new `BuiltinTemplate` enum in a template

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify