fix(walletd): correctly set indexer url via JRPC (#878)
Description
---
Fixes bug where the indexer URL is set temporarily but is not persisted
on restart

Motivation and Context
---
The indexer URL setting did not persist when set via JRPC.
Removes the indexer config from the wallet SDK. The wallet SDK should be
decoupled from the indexer. Similarly, the interface to the network used
to resolve substates does not necessarily have a JRPC endpoint to set
and so cannot be part of the interface.


How Has This Been Tested?
---

Manually

What process can a PR reviewer use to test or verify this change?
---
Setting indexer URL and checking that the setting is persisted across
webUI page refreshes

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify