feat: get all txs from pool (#888)
Description
---
Add the possibility to get all transaction in the transaction pool.

Motivation and Context
---
This is for debugging, I'm using this in the dan-testing, when the
transaction is not being processed you will see which node in which
bucket have the transaction unprocessed, why. And you will see VNs that
doesn't have the transaction in their pool showing if they know the
transaction and it was aborted or committed.

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify