feat: don't add substates that are already being processed (#885)
Description
---
Don't add transaction to a proposal if any of their substates is being
processed.

Motivation and Context
---
I left one TODO in the code. Because if we don't switch to the "no
version" transaction, them we can add aborts directly to the proposal.

How Has This Been Tested?
---
I tried 9 concurrent transaction 9 accounts to 1, and 1 accounts to 9.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify